requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==0.7.13
annotated-types==0.6.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
astroid==2.15.6
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.4
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.9.1
bleach==6.0.0
blinker==1.6.2
bs4==0.0.1
certifi==2024.2.2
cffi==1.15.1
channels==3.0.5
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudinary==1.39.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
constantly==15.1.0
contourpy==1.1.1
cookiecutter==2.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cycler==0.11.0
daphne==3.0.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
diff-match-patch==20230430
dill==0.3.7
dj-database-url==2.0.0
Django==4.2.4
django-allauth==0.54.0
django-chat==0.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.1.0
django-eventstream==4.5.1
django-grip==3.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-sse==0.4.1
django-templated-mail==1.1.1
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
docstring-to-markdown==0.12
docutils==0.20.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.13.1
fiona==1.9.5
flake8==6.0.0
Flask==3.0.0
fontawesomefree==6.5.1
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.2
geopandas==0.3.0
googledrivedownloader==0.4
googlesearch-python==1.2.3
gripcontrol==4.2.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
incremental==22.10.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
itypes==1.2.0
jaraco.classes==3.3.0
jedi==0.18.2
jellyfish==1.0.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keyring==24.2.0
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
lxml==4.9.3
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.1
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.24.2
numpydoc==1.5.0
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.2
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.18.0
plotly-geo==1.0.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pubcontrol==3.5.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydocstyle==6.3.0
pyflakes==3.0.1
pygame==2.5.2
Pygments==2.16.1
PyJWT==2.7.0
pylint==2.17.5
pylint-venv==3.0.2
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyodbc==4.0.39
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyproj==3.6.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyshp==1.2.10
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.1.1
python-lsp-server==1.7.4
python-slugify==8.0.1
python3-openid==3.2.0
pytoolconfig==1.2.5
pytz==2023.3
pywaffle==1.1.0
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
QDarkStyle==3.1
qstylizer==0.2.2
QtAwesome==1.2.3
qtconsole==5.4.4
QtPy==2.4.0
redis==4.6.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rope==1.9.0
rpds-py==0.10.3
Rtree==1.0.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
selenium==4.12.0
Send2Trash==1.8.2
service-identity==23.1.0
shapely==2.0.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.4.1
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
spyder==5.4.5
spyder-kernels==2.4.4
spyder-notebook==0.5.1
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tenacity==8.2.3
terminado==0.17.1
text-unidecode==1.3
textdistance==4.5.0
threadpoolctl==3.1.0
three-merge==0.1.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchsummary==1.5.1
torchvision==0.17.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.10.0
trio==0.22.2
trio-websocket==0.10.4
twilio==8.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.8.0
undetected-chromedriver==3.5.4
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.3
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==12.0
Werkzeug==3.0.0
whatthepatch==1.0.5
whitenoise==6.4.0
wrapt==1.15.0
wsproto==1.2.0
yapf==0.40.1
yarl==1.9.2
zipp==3.17.0
zope.interface==6.0
